tools/current_affairs/requirements.txt
python-dateutil==2.9.0.post0
firebase-admin==6.8.0
